Tilemap; fix null tile attempts to be dropped. Thanks @joekinley for finding.

+++ luxe/tilemaps/Tilemap.hx
@@ -156,7 +156,9 @@ class TilemapVisual {
 
                 for(row in layer) {
                     for(tile in row) {
-                        tile.drop();
+                        if(tile != null) {
+                            tile.drop();
+                        }
                         tile = null;
                     } //tile
                     row = null;
